chore: update .zsrc

+++ home-manager/zsh/.zshrc
@@ -8,11 +8,15 @@ export PATH="$HOME/.local/bin:$PATH"
 alias mdlint='markdownlint-cli2 "**/*.md" --config ~/.markdownlint-cli2.jsonc'
 
 function init_ts() {
+  if [ -d ".git" ]; then
+    echo ".git directory already exists. Please run this command in an empty directory."
+    return 1
+  fi
   git clone git@github.com.ph:ph0ryn/ts-base.git .
   rm -rf .git
   git init
   git add .
-  git commit -m "init"
+  git commit -m "Initial commit"
 }
 
 # npm cli
@@ -25,6 +29,9 @@ case ":$PATH:" in
   *) export PATH="$PNPM_HOME:$PATH" ;;
 esac
 # pnpm end
+
+# indexion
+export PATH="$HOME/.indexion/bin:$PATH"
 
 # ADB
 export PATH=$PATH:$HOME/Library/Android/sdk/platform-tools/
